+++ storage/azure/src/AzureClientStorage.ts
@@ -1,95 +1,12 @@
-/*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
-
-import { Readable } from "stream";
-
 import { BlockBlobClient } from "@azure/storage-blob";
-import { injectable } from "inversify";
 
-import {
-  ClientStorage,
-  instanceOfUrlDownloadInput,
-  instanceOfUrlUploadInput,
-  TransferData,
-  UrlDownloadInput,
-  UrlUploadInput,
-} from "@itwin/object-storage-core/lib/frontend";
+import { AzureFrontendStorage } from "./AzureFrontendStorage";
+import { BlockBlobClientWrapper } from "./BlockBlobClientWrapper";
 
-import { FrontendBlockBlobClientWrapper } from "./BlockBlobClientWrapper";
-import { buildBlobUrlFromConfig } from "./Helpers";
-import {
-  AzureConfigDownloadInput,
-  AzureConfigUploadInput,
-  AzureUploadInMultiplePartsInput,
-} from "./Interfaces";
-
-@injectable()
-export class AzureFrontendStorage extends ClientStorage {
-  protected getBlockBlobClientWrapper(
+export class AzureClientStorage extends AzureFrontendStorage {
+  protected override getBlockBlobClientWrapper(
     blobClient: BlockBlobClient
-  ): FrontendBlockBlobClientWrapper {
-    return new FrontendBlockBlobClientWrapper(blobClient);
-  }
-
-  public download(
-    input: (UrlDownloadInput | AzureConfigDownloadInput) & {
-      transferType: "buffer";
-    }
-  ): Promise<Buffer>;
-
-  public download(
-    input: (UrlDownloadInput | AzureConfigDownloadInput) & {
-      transferType: "stream";
-    }
-  ): Promise<Readable>;
-
-  public download(
-    input: (UrlDownloadInput | AzureConfigDownloadInput) & {
-      transferType: "local";
-      localPath: string;
-    }
-  ): Promise<string>;
-
-  public async download(
-    input: UrlDownloadInput | AzureConfigDownloadInput
-  ): Promise<TransferData> {
-    const blobClient = new BlockBlobClient(
-      instanceOfUrlDownloadInput(input)
-        ? input.url
-        : buildBlobUrlFromConfig(input)
-    );
-
-    return this.getBlockBlobClientWrapper(blobClient).download(
-      input.transferType,
-      input.localPath
-    );
-  }
-
-  public async upload(
-    input: UrlUploadInput | AzureConfigUploadInput
-  ): Promise<void> {
-    const blobClient = new BlockBlobClient(
-      instanceOfUrlUploadInput(input)
-        ? input.url
-        : buildBlobUrlFromConfig(input)
-    );
-
-    return this.getBlockBlobClientWrapper(blobClient).upload(
-      input.data,
-      input.metadata
-    );
-  }
-
-  public async uploadInMultipleParts(
-    input: AzureUploadInMultiplePartsInput
-  ): Promise<void> {
-    const blobClient = new BlockBlobClient(buildBlobUrlFromConfig(input));
-
-    return this.getBlockBlobClientWrapper(blobClient).uploadInMultipleParts(
-      input.data,
-      input.options
-    );
+  ): BlockBlobClientWrapper {
+    return new BlockBlobClientWrapper(blobClient);
   }
 }
