+++ storage/azure/src/AzureFrontendStorage.ts
@@ -1,0 +1,95 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+
+import { Readable } from "stream";
+
+import { BlockBlobClient } from "@azure/storage-blob";
+import { injectable } from "inversify";
+
+import {
+  ClientStorage,
+  instanceOfUrlDownloadInput,
+  instanceOfUrlUploadInput,
+  TransferData,
+  UrlDownloadInput,
+  UrlUploadInput,
+} from "@itwin/object-storage-core/lib/frontend";
+
+import { FrontendBlockBlobClientWrapper } from "./FrontendBlockBlobClientWrapper";
+import { buildBlobUrlFromConfig } from "./Helpers";
+import {
+  AzureConfigDownloadInput,
+  AzureConfigUploadInput,
+  AzureUploadInMultiplePartsInput,
+} from "./Interfaces";
+
+@injectable()
+export class AzureFrontendStorage extends ClientStorage {
+  protected getBlockBlobClientWrapper(
+    blobClient: BlockBlobClient
+  ): FrontendBlockBlobClientWrapper {
+    return new FrontendBlockBlobClientWrapper(blobClient);
+  }
+
+  public download(
+    input: (UrlDownloadInput | AzureConfigDownloadInput) & {
+      transferType: "buffer";
+    }
+  ): Promise<Buffer>;
+
+  public download(
+    input: (UrlDownloadInput | AzureConfigDownloadInput) & {
+      transferType: "stream";
+    }
+  ): Promise<Readable>;
+
+  public download(
+    input: (UrlDownloadInput | AzureConfigDownloadInput) & {
+      transferType: "local";
+      localPath: string;
+    }
+  ): Promise<string>;
+
+  public async download(
+    input: UrlDownloadInput | AzureConfigDownloadInput
+  ): Promise<TransferData> {
+    const blobClient = new BlockBlobClient(
+      instanceOfUrlDownloadInput(input)
+        ? input.url
+        : buildBlobUrlFromConfig(input)
+    );
+
+    return this.getBlockBlobClientWrapper(blobClient).download(
+      input.transferType,
+      input.localPath
+    );
+  }
+
+  public async upload(
+    input: UrlUploadInput | AzureConfigUploadInput
+  ): Promise<void> {
+    const blobClient = new BlockBlobClient(
+      instanceOfUrlUploadInput(input)
+        ? input.url
+        : buildBlobUrlFromConfig(input)
+    );
+
+    return this.getBlockBlobClientWrapper(blobClient).upload(
+      input.data,
+      input.metadata
+    );
+  }
+
+  public async uploadInMultipleParts(
+    input: AzureUploadInMultiplePartsInput
+  ): Promise<void> {
+    const blobClient = new BlockBlobClient(buildBlobUrlFromConfig(input));
+
+    return this.getBlockBlobClientWrapper(blobClient).uploadInMultipleParts(
+      input.data,
+      input.options
+    );
+  }
+}
