+++ storage/azure/src/index.ts
@@ -2,10 +2,10 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-export * from "./AzureClientStorage.backend";
+export * from "./AzureClientStorage";
 export * from "./AzureClientStorageBindings";
 export * from "./AzureServerStorage";
 export * from "./AzureServerStorageBindings";
-export * from "./BlockBlobClientWrapper.backend";
+export * from "./BlockBlobClientWrapper";
 export * from "./BackendHelpers";
 export * from "./frontend";
