+++ storage/s3/src/S3FrontendStorageBindings.ts
@@ -10,7 +10,7 @@ import {
   ClientStorageDependency,
 } from "@itwin/object-storage-core";
 
-import { S3ClientStorageConfig, S3FrontendStorage } from "./S3ClientStorage";
+import { S3ClientStorageConfig, S3FrontendStorage } from "./S3FrontendStorage";
 import { Types } from "./Types";
 
 export type S3ClientStorageBindingsConfig = S3ClientStorageConfig &
