+++ storage/s3/src/S3ServerStorage.ts
@@ -23,7 +23,7 @@ import {
   Types,
 } from "@itwin/object-storage-core";
 
-import { S3ClientWrapper } from "./S3ClientWrapper.backend";
+import { S3ClientWrapper } from "./S3ClientWrapper";
 
 export interface S3ServerStorageConfig {
   baseUrl: string;
